Pipeline for iris dataset
Created this pipeline for the iris dataset. It's just based on the lab videos. It compares a SVM and a DT model, and then is supposed to select the best model as the final model, but that didnt work, so now it just selected the DT model standard :). You can do endpoint and batch predictions with it, so it seems to work.